refactor(cli): extract 13 inline @cli.group() definitions into commands/ submodules
Split the monolithic cli.py (7402 lines) into focused command modules under
src/dopemux/commands/. Each group now lives in its own file with correct
relative imports (..module). cli.py reduced to ~4606 lines and uses
cli.add_command() for all groups.

New modules: autoresponder, capture_group, code, decisions, dev, extract,
extractor, memory, profile, trigger_group, update, upgrades, workflow_group.

Also: add Serena ~/.serena and host code volume mounts to compose.yml,
remove stale docker/mcp-servers-source/docker-compose.yml duplicate,
update test monkeypatches to target capture/trigger module namespaces.
