+++ compose.yml
@@ -509,6 +509,8 @@ services:
       - HOST_PROJECT_RELATIVE_PATH=${HOST_PROJECT_RELATIVE_PATH}
     volumes:
       - ${HOST_CODE_PARENT_DIR:-/tmp}:/workspaces
+      - /Users/hue/.serena:/root/.serena:rw
+      - /Users/hue/code:/Users/hue/code:ro
     healthcheck:
       test: ["CMD-SHELL", "curl -f http://localhost:4006/health || exit 1"]
       timeout: 10s
